Moved and renamed decorator for figure tests
